fix editor, remove wpf and improve voxel

+++ library/modules/runtime/runtimeshared/include/tiki/runtimeshared/freecamera.hpp
@@ -23,7 +23,7 @@ namespace tiki
 		bool		getMouseControl() const			{ return m_enableMouse; }
 		void		setMouseControl( bool enabled );
 
-		void		update( Camera& targetCamera, float timeDelta );
+		void		update( Camera& targetCamera, double timeDelta );
 
 		bool		processInputEvent( const InputEvent& inputEvent );
